requirements.txt
opencv-python-headless
PyQt5
pillow
pytesseract
numpy
sympy
pygments